+++ language/duchain/definitions.cpp
@@ -76,7 +76,7 @@ class DefinitionsRequestItem {
     return m_item.hash();
   }
   
-  size_t itemSize() const {
+  uint itemSize() const {
       return m_item.itemSize();
   }
